backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic[email]
httpx==0.27.0
supabase==2.4.4
opencv-python-headless==4.9.0.80
ultralytics==8.1.28
easyocr==1.7.1
numpy==1.26.4
pillow==10.2.0
pandas==2.2.1
python-dateutil==2.9.0.post0
redis==5.0.3
orjson==3.10.0
loguru==0.7.2
insightface==0.7.3
onnxruntime-gpu==1.23.2
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
bcrypt==4.1.2
